Simple NBT Logic
This would force Mods that want to register over NBT to set the NBT tag
bettercombatparent to the same value as the Parent in the Json would be set to.
This kind of requires to parse WeaponConfigurations of Weapons that dont exist so the Parent can be found and accessed properly.